Disassembler: Add offset to the block's PC in struct opcode
This will later permit to remove opcodes from the IR, since we won't
calculate the opcode's PC according to the number of opcodes preceding
it.

+++ lightrec-private.h
@@ -107,7 +107,7 @@ struct lightrec_state {
 };
 
 u32 lightrec_rw(struct lightrec_state *state, union code op,
-		u32 addr, u32 data, u32 *flags);
+		u32 addr, u32 data, u16 *flags);
 
 void lightrec_free_block(struct block *block);
